Started the new config handling
Implemented new and not complete config handling and implemented command parameter support
