+++ old_src/config.ini
@@ -1,0 +1,26 @@
+[General]
+; = If a local plugin folder exists (True/False) (If False SFTP/FTP will be used):
+uselocalpluginfolder = True
+
+[Local - This Machine]
+pathtopluginfolder = C:/Users/USER/Desktop/plugins
+; = For a different folder to store the updated plugins change to (True/False) and the path below
+seperatedownloadpath = False
+pathtoseperatedownloadpath = C:/Users/USER/Desktop/plugins
+
+[SFTP - Remote Server]
+server = 0.0.0.0
+username = user
+password = password
+; = If a different Port for SFTP needs to be used (Default: 22)
+sftpport = 22
+;_ = If a different Port for FTP needs to be used (Default: 21)
+ftpport = 21
+;__ = Change the path below if the plugin folder path is different on the SFTP/FTP server (Change only if you know what you are doing)
+pluginfolderonserver = /plugins
+;___ = If you want to use FTP instead of SFTP change to (False) else use (True)
+use_sftp = True
+;____ = For a different folder to store the updated plugins (Only with the update command!) change to (True/False) and the path below
+seperatedownloadpath = False
+pathtoseperatedownloadpath = /plugins
+
