+++ old_src/plugin/plugin_remover.py
